+++ packages/core-data/src/entities.js
@@ -19,6 +19,18 @@ export const DEFAULT_ENTITY_KEY = 'id';
 
 const POST_RAW_ATTRIBUTES = [ 'title', 'excerpt', 'content' ];
 
+// A hardcoded list of post types that support revisions.
+// Reflects post types in Core's src/wp-includes/post.php.
+// @TODO: Ideally this should be fetched from the  `/types` REST API's view context.
+const POST_TYPE_ENTITIES_WITH_REVISIONS_SUPPORT = [
+	'post',
+	'page',
+	'wp_block',
+	'wp_navigation',
+	'wp_template',
+	'wp_template_part',
+];
+
 export const rootEntitiesConfig = [
 	{
 		label: __( 'Base' ),
@@ -205,8 +217,16 @@ export const rootEntitiesConfig = [
 		kind: 'root',
 		baseURL: '/wp/v2/global-styles',
 		baseURLParams: { context: 'edit' },
-		plural: 'globalStylesVariations', // Should be different than name.
+		plural: 'globalStylesVariations', // Should be different from name.
 		getTitle: ( record ) => record?.title?.rendered || record?.title,
+		getRevisionsUrl: ( parentId, revisionId ) =>
+			`/wp/v2/global-styles/${ parentId }/revisions${
+				revisionId ? '/' + revisionId : ''
+			}`,
+		supports: {
+			revisions: true,
+		},
+		supportsPagination: true,
 	},
 	{
 		label: __( 'Themes' ),
@@ -295,6 +315,11 @@ async function loadPostTypeEntities() {
 				selection: true,
 			},
 			mergedEdits: { meta: true },
+			supports: {
+				revisions: POST_TYPE_ENTITIES_WITH_REVISIONS_SUPPORT.includes(
+					postType?.slug
+				),
+			},
 			rawAttributes: POST_RAW_ATTRIBUTES,
 			getTitle: ( record ) =>
 				record?.title?.rendered ||
@@ -328,6 +353,13 @@ async function loadPostTypeEntities() {
 			syncObjectType: 'postType/' + postType.name,
 			getSyncObjectId: ( id ) => id,
 			supportsPagination: true,
+			getRevisionsUrl: ( parentId, revisionId ) =>
+				`/${ namespace }/${
+					postType.rest_base
+				}/${ parentId }/revisions${
+					revisionId ? '/' + revisionId : ''
+				}`,
+			revisionKey: isTemplate ? 'wp_id' : DEFAULT_ENTITY_KEY,
 		};
 	} );
 }