+++ packages/e2e-tests/plugins/interactive-blocks/directive-context/render.php
@@ -5,14 +5,16 @@
  * @package gutenberg-test-interactive-blocks
  */
 
+gutenberg_enqueue_module( 'directive-context-view' );
 ?>
-<div data-wp-interactive>
+
+<div data-wp-interactive='{"namespace": "directive-context"}'>
 	<div
 		data-wp-context='{ "prop1":"parent","prop2":"parent","obj":{"prop4":"parent","prop5":"parent"},"array":[1,2,3] }'
 	>
 		<pre
 			data-testid="parent context"
-			data-wp-bind--children="derived.renderContext"
+			data-wp-bind--children="state.renderContext"
 		>
 			<!-- rendered during hydration -->
 		</pre>
@@ -53,7 +55,7 @@
 		>
 			<pre
 				data-testid="child context"
-				data-wp-bind--children="derived.renderContext"
+				data-wp-bind--children="state.renderContext"
 			>
 				<!-- rendered during hydration -->
 			</pre>
@@ -121,7 +123,7 @@
 </div>
 
 <div
-	data-wp-interactive
+	data-wp-interactive='{"namespace": "directive-context-navigate"}'
 	data-wp-navigation-id="navigation"
 	data-wp-context='{ "text": "first page" }'
 >