+++ packages/react-native-editor/CHANGELOG.md
@@ -10,7 +10,11 @@ For each user feature we should also add a importance categorization label  to i
 -->
 
 ## Unreleased
+-   [***] Fix issue when backspacing in an empty Paragraph block [#56496]
+
+## 1.109.0
 -   [*] Audio block: Improve legibility of audio file details on various background colors [#55627]
+-   [*] In the deeply nested block warning, only display the ungroup option for blocks that support it [#56445]
 
 ## 1.108.0
 -   [*] Fix error when pasting deeply nested structure content [#55613]